requirements.txt
tensorflow>=2.11
streamlit
pillow
numpy
